pyproject.toml
[project]
name = "ml-comp-framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "albumentations>=2.0.8",
    "hydra-core>=0.22.0",
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "scikit-learn>=1.6.1",
    "torch",
    "torchvision",
    "wandb",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
